# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hunches - Prediction platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  icon: string;
  hunchCount: number;
  color?: string;
}

export type PrizeType = typeof PrizeType[keyof typeof PrizeType];


export const PrizeType = {
  gift_card: 'gift_card',
  merch: 'merch',
  cash_equivalent: 'cash_equivalent',
} as const;

export interface Prize {
  id: number;
  label: string;
  type: PrizeType;
  value: string;
  /** @nullable */
  imageUrl?: string | null;
}

export type HunchStatus = typeof HunchStatus[keyof typeof HunchStatus];


export const HunchStatus = {
  open: 'open',
  closed: 'closed',
  resolved: 'resolved',
} as const;

export interface HunchOption {
  id: number;
  label: string;
  percentage: number;
}

export interface Hunch {
  id: number;
  title: string;
  description: string;
  categorySlug: string;
  categoryName: string;
  categoryColor?: string;
  status: HunchStatus;
  participantCount: number;
  endsAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  prize: Prize;
  options: HunchOption[];
  featured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  winnerOption?: string | null;
}

export interface HunchList {
  hunches: Hunch[];
  total: number;
}

export interface PlatformStats {
  totalHunches: number;
  totalParticipants: number;
  totalPrizesAwarded: number;
  activeHunches: number;
}

export interface PredictionInput {
  optionId: number;
}

export interface Prediction {
  id: number;
  hunchId: number;
  optionId: number;
  optionLabel: string;
  createdAt: string;
}

export type ListHunchesParams = {
/**
 * Filter by category slug
 */
category?: string;
/**
 * Filter by status
 */
status?: ListHunchesStatus;
/**
 * Filter featured hunches only
 */
featured?: boolean;
limit?: number;
offset?: number;
};

export type ListHunchesStatus = typeof ListHunchesStatus[keyof typeof ListHunchesStatus];


export const ListHunchesStatus = {
  open: 'open',
  closed: 'closed',
  resolved: 'resolved',
} as const;