# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hunches - Prediction platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  slug: string;
  name: string;
  icon: string;
  hunchCount: number;
  color?: string;
}