# lib/api-zod/src/generated/types/hunch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hunches - Prediction platform API
 * OpenAPI spec version: 0.1.0
 */
import type { HunchOption } from './hunchOption';
import type { HunchStatus } from './hunchStatus';
import type { Prize } from './prize';

export interface Hunch {
  id: number;
  title: string;
  description: string;
  categorySlug: string;
  categoryName: string;
  categoryColor?: string;
  status: HunchStatus;
  participantCount: number;
  endsAt: Date;
  /** @nullable */
  resolvedAt?: string | null;
  prize: Prize;
  options: HunchOption[];
  featured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  winnerOption?: string | null;
}