# lib/api-zod/src/generated/types/hunchList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hunches - Prediction platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Hunch } from './hunch';

export interface HunchList {
  hunches: Hunch[];
  total: number;
}