# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hunches - Prediction platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalHunches: number;
  totalParticipants: number;
  totalPrizesAwarded: number;
  activeHunches: number;
}