# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hunches - Prediction platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Prediction {
  id: number;
  hunchId: number;
  optionId: number;
  optionLabel: string;
  createdAt: Date;
}