# lib/api-zod/src/generated/types/prize.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hunches - Prediction platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PrizeType } from './prizeType';

export interface Prize {
  id: number;
  label: string;
  type: PrizeType;
  value: string;
  /** @nullable */
  imageUrl?: string | null;
}